move faker overrides into miniLcm tests
