--- backend/FwLite/FwLiteProjectSync.Tests/Fixtures/WritingSystemCodes.cs
@@ -1,8 +1,0 @@
-using SIL.WritingSystems;
-
-namespace FwLiteProjectSync.Tests;
-
-public static class WritingSystemCodes
-{
-    public static string[] ValidTwoLetterCodes = StandardSubtags.RegisteredLanguages.Select(lang => lang.Code).ToArray();
-}
