+++ backend/FwLite/LcmCrdt.Tests/EntityCopyMethodTests.cs
@@ -1,5 +1,6 @@
 ﻿using SIL.Harmony.Entities;
 using LcmCrdt.Objects;
+using MiniLcm.Tests.FakerOverrids;
 using SIL.Harmony;
 using Soenneker.Utils.AutoBogus;
 using Soenneker.Utils.AutoBogus.Config;
