+++ backend/FwLite/MiniLcm.Tests/FakerOverrids/MultiStringOverride.cs
@@ -1,8 +1,8 @@
-using MiniLcm.Models;
+﻿using MiniLcm.Models;
 using Soenneker.Utils.AutoBogus.Context;
 using Soenneker.Utils.AutoBogus.Override;
 
-namespace FwLiteProjectSync.Tests.Fixtures;
+namespace MiniLcm.Tests.FakerOverrids;
 
 public class MultiStringOverride: AutoFakerOverride<MultiString>
 {
@@ -16,7 +16,7 @@ public class MultiStringOverride: AutoFakerOverride<MultiString>
         var wordsArray = context.Faker.Random.WordsArray(1, 4);
         foreach (var word in wordsArray)
         {
-            var ws = context.Faker.Random.ArrayElement(MiniLcm.Tests.WritingSystemCodes.ValidTwoLetterCodes);
+            var ws = context.Faker.Random.ArrayElement(WritingSystemCodes.ValidTwoLetterCodes);
             target[ws] = word;
         }
     }
